refactor: renamed module `workbench_tao` -> `workbench_desktop`
