+++ internal/workbench/desktop/src/contributions.rs
