+++ internal/workbench/desktop/src/contributions/tree_view_groups.rs
