+++ internal/workbench/desktop/src/parts/primary_activitybar.rs
