+++ internal/workbench/desktop/src/parts/primary_sidebar.rs
