+++ internal/workbench/desktop/src/view.rs
