+++ view/desktop/bin/Cargo.toml
@@ -32,7 +32,7 @@ platform_configuration.workspace = true
 platform_formation.workspace = true
 platform_fs.workspace = true
 
-workbench_tao.workspace = true
+workbench_desktop.workspace = true
 workbench_service_configuration_tao.workspace = true
 workbench_service_environment_tao.workspace = true
