+++ view/desktop/bin/src/command/cmd_base.rs
@@ -1,11 +1,13 @@
 use tauri::State;
-use workbench_tao::contributions::resents::{RecentsViewContentProviderOutput, RecentsViewModel};
-use workbench_tao::parts::primary_activitybar::{
+use workbench_desktop::contributions::resents::{
+    RecentsViewContentProviderOutput, RecentsViewModel,
+};
+use workbench_desktop::parts::primary_activitybar::{
     DescribeActivityBarPartOutput, PrimaryActivityBarPart,
 };
-use workbench_tao::parts::primary_sidebar::{DescribeSideBarPartOutput, PrimarySideBarPart};
-use workbench_tao::parts::{AnyPart, Parts};
-use workbench_tao::window::NativePlatformInfo;
+use workbench_desktop::parts::primary_sidebar::{DescribeSideBarPartOutput, PrimarySideBarPart};
+use workbench_desktop::parts::{AnyPart, Parts};
+use workbench_desktop::window::NativePlatformInfo;
 
 use crate::AppState;
