+++ view/desktop/bin/src/command/cmd_dummy.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use platform_core::context_v2::async_context::AsyncContext;
 use std::path::PathBuf;
 use tauri::{AppHandle, Manager, State};
-use workbench_tao::WorkbenchState;
+use workbench_desktop::WorkbenchState;
 
 use crate::AppState;
